+++ app/views/hareruya_freee/partners/index.html.haml
@@ -1,0 +1,3 @@
+- @partners.each do |partner|
+  %div
+    = link_to partner.name, partner_path(partner)
