--- app/views/layouts/the245plans_freee/application.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <title>The245plansFreee</title>
-  <%= stylesheet_link_tag    "the245plans_freee/application", media: "all" %>
-  <%= javascript_include_tag "the245plans_freee/application" %>
-  <%= csrf_meta_tags %>
-</head>
-<body>
-
-<%= yield %>
-
-</body>
-</html>
