--- app/views/the245plans_freee/partners/index.html.haml
@@ -1,3 +1,0 @@
-- @partners.each do |partner|
-  %div
-    =partner.name
