+++ config/routes.rb
@@ -1,5 +1,5 @@
-The245plansFreee::Engine.routes.draw do
-  resources :partners, only: :index
+HareruyaFreee::Engine.routes.draw do
+  resources :partners, only: [:index, :show]
   get 'refresh_token' => 'refresh_token#index'
   get 'callback' => 'refresh_token#callback'
 end
