+++ db/migrate/20150721123840_rename_from245plans_to_hareruya.rb
@@ -1,0 +1,5 @@
+class RenameFrom245plansToHareruya < ActiveRecord::Migration
+  def change
+    rename_table :the245plans_freee_partners, :hareruya_freee_partners
+  end
+end
