+++ lib/hareruya_freee.rb
@@ -1,0 +1,4 @@
+require "hareruya_freee/engine"
+
+module HareruyaFreee
+end
