+++ lib/hareruya_freee/engine.rb
@@ -1,0 +1,5 @@
+module HareruyaFreee
+  class Engine < ::Rails::Engine
+    isolate_namespace HareruyaFreee
+  end
+end
