+++ lib/hareruya_freee/version.rb
@@ -1,3 +1,3 @@
-module The245plansFreee
+module HareruyaFreee
   VERSION = "0.0.1"
 end
