+++ lib/tasks/hareruya_freee_tasks.rake
@@ -1,4 +1,4 @@
 # desc "Explaining what the task does"
-# task :the245plans_freee do
+# task :hareruya_freee do
 #   # Task goes here
 # end
