--- lib/the245plans_freee.rb
@@ -1,4 +1,0 @@
-require "the245plans_freee/engine"
-
-module The245plansFreee
-end
