--- lib/the245plans_freee/engine.rb
@@ -1,5 +1,0 @@
-module The245plansFreee
-  class Engine < ::Rails::Engine
-    isolate_namespace The245plansFreee
-  end
-end
