+++ test/dummy/config/routes.rb
@@ -1,4 +1,4 @@
 Rails.application.routes.draw do
 
-  mount The245plansFreee::Engine => "/the245plans_freee"
+  mount HareruyaFreee::Engine => "/hareruya_freee"
 end
