+++ test/hareruya_freee_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class HareruyaFreeeTest < ActiveSupport::TestCase
+  test "truth" do
+    assert_kind_of Module, HareruyaFreee
+  end
+end
