--- test/the245plans_freee_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-class The245plansFreeeTest < ActiveSupport::TestCase
-  test "truth" do
-    assert_kind_of Module, The245plansFreee
-  end
-end
